Feat: Implemented Snake Body movement I have used the old MySnake class, along with some of Paddle to properly implement a Model-Controller behaviour for the Snake. The snake moves at an editable frame rate, at a min of 1 ms frame times, and it replicates the old swing implementation completely.
As this is a basic implementation, and without food yet, I have hard
coded a test of 2 snake body instances. Ideally, I do not want the
Snake controller creating GUI elements, so I will move that to a View.

There are no explicit views in my game as of yet, and due to its game
nature gui elements are separated in factories and created through them.
So, as a compromise, I will create a View for the general 'Snake',
which creates all the snake gui elements. Atleast, that is the plan.