Refactor: Partial implementation of Snake View As I am using FXGL, the use of Views and whatnot becomes rather strained in its architecture, but it is plausible to make a somewhat loose definition on what a controller, view and model is. The model is the easiest to implement, and for some aspects I can use a FXML for game menus, so a View can be unproblematic.
However, a View for the Entities of the game, such as the Snake and its
various moving body parts, is not so easy to explicitly define. I have
decided to go with the method of having a 'SnakeView' class that handles
creation and storage of the Snake and its body entities, which are
essentially JavaFX nodes.

This follows the paradigm of View storing only application GUIs, moreso
what you can see, and here the View stores the main playing snake.

This commit is partially completed. I will finish it tomorrow as moving
around this architecture is rather tiresome.