Feat: Added Level support. As FXGL supports .tmx level files, I have used a tile-based level system that uses level files in the resources folder to build obstacles.
These levels only needed to be loaded, and all the obstacles within them
are automatically placed, and their collision functionality enabled. I
have made 3 levels to correspond to 3 difficulties, which the user
selects on new game. This difficulty is also saved on the high-Score
leaderboard.

This system allows high extensibility, as I have also made it as easy
as possible to add new levels and new difficulties to the game in the
'Config' class.